Create Deployment.md
#89

Your README.md should contain a description of your project along the following lines
An outline of the application you were aiming to build, target users, data sources etc (similar to the proposal)
A reference to your second Github repository if you used one
A description of what you have been able to implement in this MVP, use your milestones to highlight what you've achieved
A guide to the project source code - where should we look for what you have done
A summary of what your next steps would be if you were to continue the project
A summary of the main roles and contributions of each team member and how you managed interaction and communication through the project

A file DEPLOYMENT.md with the following details
A link to a deployed version of your application if available
Information on how to run and build the project, eg. if someone were to take over development
Information about any use of Continuous Integration you have implemented.
A folder SCREENSHOTS containing screenshots (or maybe a video) of key pages in your application and a README.md file that describes what they show.**